# api/openapi.yaml
components:
  schemas:
...
    Create-api-keyRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/Create-api-keyRequest.json
          format: uri
          readOnly: true
          type: string
        capabilities:
          items:
            type: string
          maxItems: 7
          minItems: 1
          type:
            - array
            - "null"
        id:
          maxLength: 63
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          type: string
      required:
        - id
        - capabilities
      type: object
...
    CreateKeyResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/CreateKeyResponse.json
          format: uri
          readOnly: true
          type: string
        key:
          $ref: "#/components/schemas/APIKey"
        secret:
          description: Shown once. Store it securely; Manifold only persists an Argon2id hash.
          type: string
      required:
        - key
        - secret
      type: object
...
    DocumentAccepted:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/DocumentAccepted.json
          format: uri
          readOnly: true
          type: string
        document:
          $ref: "#/components/schemas/Document"
        folders_created:
          items:
            $ref: "#/components/schemas/Folder"
          type:
            - array
            - "null"
        job:
          $ref: "#/components/schemas/Job"
      required:
        - document
        - job
      type: object
...
    PageAPIKey:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/PageAPIKey.json
          format: uri
          readOnly: true
          type: string
        items:
          items:
            $ref: "#/components/schemas/APIKey"
          type:
            - array
            - "null"
        next_cursor:
          type: string
      required:
        - items
      type: object
...
    PageDocument:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/PageDocument.json
          format: uri
          readOnly: true
          type: string
        items:
          items:
            $ref: "#/components/schemas/Document"
          type:
            - array
            - "null"
        next_cursor:
          type: string
      required:
        - items
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: Manifold API key
      description: A Manifold API key with the capabilities required by the operation.
      scheme: bearer
      type: http
info:
  description: A self-hosted knowledge and memory layer for AI agents.
  title: Manifold API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/v1/api-keys:
    get:
      operationId: list-api-keys
      parameters:
        - explode: false
          in: query
          name: cursor
          schema:
            pattern: ^c[0-9a-z]+$
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 50
            format: int64
            maximum: 200
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAPIKey"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Request Entity Too Large
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Service Unavailable
      summary: List API-key records
      tags:
        - Authentication
    post:
      operationId: create-api-key
      parameters:
        - description: A caller-generated key that makes this mutation safe to retry.
          in: header
          name: Idempotency-Key
          schema:
            description: A caller-generated key that makes this mutation safe to retry.
            maxLength: 128
            minLength: 8
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Create-api-keyRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateKeyResponse"
          description: Created
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Request Entity Too Large
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Service Unavailable
      summary: Create an API key
      tags:
        - Authentication
...
  /api/v1/documents:
    get:
      operationId: list-documents
      parameters:
        - explode: false
          in: query
          name: folder_id
          schema:
            pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
            type: string
        - explode: false
          in: query
          name: cursor
          schema:
            pattern: ^c[0-9a-z]+$
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 50
            format: int64
            maximum: 200
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDocument"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Request Entity Too Large
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Service Unavailable
      summary: List active documents
      tags:
        - Documents
    post:
      operationId: create-document
      parameters:
        - description: A caller-generated key that makes this mutation safe to retry.
          in: header
          name: Idempotency-Key
          schema:
            description: A caller-generated key that makes this mutation safe to retry.
            maxLength: 128
            minLength: 8
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequestBody"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAccepted"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Request Entity Too Large
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Service Unavailable
      summary: Create and asynchronously index a document
      tags:
        - Documents
...
  /api/v1/ui/session:
    get:
      operationId: get-ui-session
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Request Entity Too Large
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Service Unavailable
      summary: Inspect the active browser session
      tags:
        - Authentication
    post:
      operationId: create-ui-session
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Create-ui-sessionRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Create-ui-sessionResponse"
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security: []
      summary: Exchange an API key for a browser session
      tags:
        - Authentication
...
security:
  - bearerAuth: []